requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
attrs=19.1.0=pypi_0
backcall=0.1.0=pypi_0
blas=1.0=mkl
bleach=3.1.0=pypi_0
ca-certificates=2019.5.15=1
certifi=2019.6.16=py36_1
cffi=1.12.3=py36h2e261b9_0
cuda100=1.0=0
cudatoolkit=10.0.130=0
cycler=0.10.0=pypi_0
decorator=4.4.0=pypi_0
defusedxml=0.6.0=pypi_0
entrypoints=0.3=pypi_0
freetype=2.9.1=h8a8886c_1
h5py=2.9.0=pypi_0
imageio=2.5.0=pypi_0
intel-openmp=2019.4=243
ipykernel=5.1.2=pypi_0
ipython=7.7.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
isodate=0.6.0=pypi_0
jedi=0.15.1=pypi_0
jinja2=2.10.1=pypi_0
joblib=0.13.2=pypi_0
jpeg=9b=h024ee3a_2
jsonschema=3.0.2=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=5.3.1=pypi_0
jupyter-console=6.0.0=pypi_0
jupyter-core=4.5.0=pypi_0
kiwisolver=1.1.0=pypi_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
markupsafe=1.1.1=pypi_0
matplotlib=3.1.1=pypi_0
mistune=0.8.4=pypi_0
mkl=2019.4=243
mkl-service=2.0.2=py36h7b6447c_0
mkl_fft=1.0.14=py36ha843d7b_0
mkl_random=1.0.2=py36hd81dba3_0
nbconvert=5.6.0=pypi_0
nbformat=4.4.0=pypi_0
ncurses=6.1=he6710b0_1
networkx=2.3=pypi_0
ninja=1.9.0=py36hfd86e86_0
notebook=6.0.0=pypi_0
numpy=1.16.4=py36h7e9f1db_0
numpy-base=1.16.4=py36hde5b4d6_0
olefile=0.46=py36_0
opencv-python=4.1.0.25=pypi_0
openssl=1.1.1c=h7b6447c_1
pandas=0.25.0=pypi_0
pandocfilters=1.4.2=pypi_0
parso=0.5.1=pypi_0
pexpect=4.7.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=6.1.0=py36h34e0f95_0
pip=19.2.2=py36_0
plyfile=0.7=pypi_0
prometheus-client=0.7.1=pypi_0
prompt-toolkit=2.0.9=pypi_0
ptyprocess=0.6.0=pypi_0
pycparser=2.19=py36_0
pygments=2.4.2=pypi_0
pyparsing=2.4.2=pypi_0
pyrsistent=0.15.4=pypi_0
python=3.6.9=h265db76_0
python-dateutil=2.8.0=pypi_0
pytz=2019.2=pypi_0
pywavelets=1.0.3=pypi_0
pyzmq=18.1.0=pypi_0
qtconsole=4.5.3=pypi_0
rdflib=4.2.2=pypi_0
readline=7.0=h7b6447c_5
scikit-image=0.15.0=pypi_0
scikit-learn=0.21.3=pypi_0
scipy=1.3.1=py36h7c811a0_0
send2trash=1.5.0=pypi_0
setuptools=41.1.0=pypi_0
six=1.12.0=pypi_0
sqlite=3.29.0=h7b6447c_0
terminado=0.8.2=pypi_0
testpath=0.4.2=pypi_0
tk=8.6.8=hbc83047_0
torch-cluster=1.4.4=pypi_0
torch-geometric=1.3.0=pypi_0
torch-scatter=1.3.1=pypi_0
torch-sparse=0.4.0=pypi_0
tornado=6.0.3=pypi_0
tqdm=4.34.0=pypi_0
traitlets=4.3.2=pypi_0
wcwidth=0.1.7=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.33.4=py36_0
widgetsnbextension=3.5.1=pypi_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
